+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/exporters/shape/SVGShapeExporter.java
@@ -135,7 +135,7 @@ public class SVGShapeExporter extends DefaultSVGShapeExporter {
     }
 
     @Override
-    public void lineStyle(double thickness, RGB color, boolean pixelHinting, String scaleMode, int startCaps, int endCaps, int joints, int miterLimit) {
+    public void lineStyle(double thickness, RGB color, boolean pixelHinting, String scaleMode, int startCaps, int endCaps, int joints, float miterLimit) {
         finalizePath();
         thickness *= zoom / SWF.unitDivisor;
         path.setAttribute("fill", "none");
@@ -169,8 +169,8 @@ public class SVGShapeExporter extends DefaultSVGShapeExporter {
                 break;
             default:
                 path.setAttribute("stroke-linejoin", "miter");
-                if (miterLimit >= 1 && miterLimit != 4) {
-                    path.setAttribute("stroke-miterlimit", Integer.toString(miterLimit));
+                if (miterLimit >= 1 && miterLimit != 4f) {
+                    path.setAttribute("stroke-miterlimit", Double.toString(miterLimit));
                 }
                 break;
         }
